Move bot service models
