+++ botx/__init__.py
@@ -12,8 +12,6 @@ from botx.bot.exceptions import (
     UnknownBotAccountError,
 )
 from botx.bot.handler_collector import HandlerCollector
-from botx.bot.models.bot_account import BotAccount
-from botx.bot.models.method_callbacks import BotAPIMethodFailedCallback
 from botx.bot.testing import lifespan_wrapper
 from botx.client.chats_api.exceptions import (
     CantUpdatePersonalChatError,
@@ -45,6 +43,7 @@ from botx.client.notifications_api.exceptions import (
 from botx.client.notifications_api.markup import BubbleMarkup, Button, KeyboardMarkup
 from botx.client.users_api.exceptions import UserNotFoundError
 from botx.models.attachments import OutgoingAttachment
+from botx.models.bot_account import BotAccount
 from botx.models.chat import Chat
 from botx.models.enums import ChatTypes, ClientPlatforms, MentionTypes, UserKinds
 from botx.models.message.entities import Mention
@@ -54,6 +53,7 @@ from botx.models.message.incoming_message import (
     UserDevice,
     UserEventSender,
 )
+from botx.models.method_callbacks import BotAPIMethodFailedCallback
 from botx.models.status import BotMenu, StatusRecipient
 from botx.models.system_events.added_to_chat import AddedToChatEvent
 from botx.models.system_events.chat_created import ChatCreatedEvent, ChatCreatedMember
